deploy/requirements.txt
numpy
pandas
faiss-cpu
scikit-learn
sentence-transformers
rank-bm25
